brahma_hq_extractor/requirements.txt
pdfplumber
fastapi==0.115.0
uvicorn[standard]==0.30.6
google-cloud-storage==2.17.0
pdfplumber
python-multipart
